Split BCSD regridding and merging with truth into two python files.
Also changes:
- Convert temperature unit from Celsius to Kelvin.
- Ignore BCSD time coords while merging with truth, and use time & grid from root to avoid inconsistency.
- Add utils.py to load different paths for local testing & BIG.
