+++ src/nc_helpers/split_n_convert_nc.py
