+++ src/nc_helpers/ssp_128x96_grid_coords.nc
